+++ test/functional/actualVsExpected.js
@@ -5,127 +5,123 @@ const { logMochaOutput, getMochaPath } = require('../testHelpers');
 const internalMochaPath = getMochaPath();
 const path = require('path');
 
-describe('Check TeamCity Output is correct with recordHookFailures option', function () {
-	let teamCityStdout, teamCityStderr, teamCityOutputArray;
+describe('Check TeamCity Output is correct with actualVsExpected option', function () {
+	let teamCityStdout, teamCityStderr, teamCityOutputArray, teamCityErrorOutputArray;
 	function verifyResults() {
 		it('stdout output should exist', function () {
 			assert.isOk(teamCityStdout, 'has output');
 			assert.isOk(teamCityOutputArray, 'array of output is populated');
-			assert.lengthOf(teamCityOutputArray, 13);
-			assert.isEmpty(teamCityOutputArray[12]);
+			assert.isOk(teamCityOutputArray.length >= 10, 'at least 10 lines of output');
+			assert.lengthOf(teamCityOutputArray, 10);
 		});
 
-		it('stderr output should not exist', function () {
-			assert.isEmpty(teamCityStderr);
+		it('stderr output should exist', function () {
+			assert.isNotOk(teamCityStderr);
+			assert.isOk(teamCityStderr.length === 0);
+			assert.lengthOf(teamCityErrorOutputArray, 1);
 		});
 
 
 		it('stdout output should exist', function () {
 			assert.isOk(teamCityStdout);
 			assert.isOk(teamCityOutputArray);
-			assert.isOk(teamCityOutputArray.length >= 12);
+			assert.isOk(teamCityOutputArray.length >= 9);
 		});
 
-		it('Suite1 started is OK', function () {
+		it('Suite started is OK', function () {
 			const rowToCheck = teamCityOutputArray[0];
 			assert.isOk(/##teamcity\[testSuiteStarted/.test(rowToCheck));
-			assert.isOk(/name='Hook Test Top Describe Fail'/.test(rowToCheck));
+			assert.isOk(/name='Top Describe'/.test(rowToCheck));
 			assert.isOk(/flowId=/.test(rowToCheck));
 			assert.isOk(/]/.test(rowToCheck));
 		});
 
-		it('Hook Test started is OK', function () {
+		it('Test started is OK', function () {
 			const rowToCheck = teamCityOutputArray[1];
 			assert.match(rowToCheck, /##teamcity\[testStarted/);
-			assert.match(rowToCheck, /name='"before all" hook'/);
+			assert.match(rowToCheck, /name='Passing Test @pass'/);
 			assert.match(rowToCheck, /flowId=/);
 			assert.match(rowToCheck, /]/);
 		});
 
-		it('Hook Test Failed is OK', function () {
+		it('Passing Test Finished is OK', function () {
 			const rowToCheck = teamCityOutputArray[2];
-			assert.isOk(/##teamcity\[testFailed/.test(rowToCheck));
-			assert.isOk(/name='"before all" hook/.test(rowToCheck));
-			assert.isOk(/details='/.test(rowToCheck));
-			assert.isOk(/Error: Before hook error fail/.test(rowToCheck));
-			assert.isOk(/|n/.test(rowToCheck));
-			assert.isOk(/|failingHook.js:29:12/.test(rowToCheck));
-			assert.isOk(/captureStandardOutput='true'/.test(rowToCheck));
+			assert.isOk(/##teamcity\[testFinished/.test(rowToCheck));
+			assert.isOk(/name='Passing Test @pass'/.test(rowToCheck));
+			assert.isOk(/flowId=/.test(rowToCheck));
+			assert.isOk(/duration=/.test(rowToCheck));
+			assert.isOk(/]/.test(rowToCheck));
+		});
+
+		it('Test Failed Started is OK', function () {
+			const rowToCheck = teamCityOutputArray[3];
+			assert.isOk(/##teamcity\[testStarted/.test(rowToCheck));
+			assert.isOk(/name='Failing Test @fail'/.test(rowToCheck));
 			assert.isOk(/flowId=/.test(rowToCheck));
 			assert.isOk(/duration=/.test(rowToCheck) === false);
 			assert.isOk(/]/.test(rowToCheck));
 		});
 
-		it('Suite1 Finished is OK', function () {
-			const rowToCheck = teamCityOutputArray[3];
-			assert.isOk(/##teamcity\[testSuiteFinished/.test(rowToCheck));
-			assert.isOk(/name='Hook Test Top Describe Fail'/.test(rowToCheck));
-			assert.isOk(/duration=/.test(rowToCheck));
-			assert.isOk(/flowId=/.test(rowToCheck));
-			assert.isOk(/]/.test(rowToCheck));
-		});
-
-		it('Suite2 started is OK', function () {
-			const rowToCheck = teamCityOutputArray[4];
-			assert.isOk(/##teamcity\[testSuiteStarted/.test(rowToCheck));
-			assert.isOk(/name='Hook Test Top Describe Pass'/.test(rowToCheck));
-			assert.isOk(/flowId=/.test(rowToCheck));
-			assert.isOk(/]/.test(rowToCheck));
-		});
-
-
-		it('Test started is OK', function () {
-			const rowToCheck = teamCityOutputArray[9];
-			assert.isOk(/##teamcity\[testStarted/.test(rowToCheck));
-			assert.isOk(/name='Test Passing Test @pass'/.test(rowToCheck));
-			assert.isOk(/flowId=/.test(rowToCheck));
-			assert.isOk(/]/.test(rowToCheck));
-		});
-
-
-		it('Passing Test Finished is OK', function () {
-			const rowToCheck = teamCityOutputArray[10];
-			assert.isOk(/##teamcity\[testFinished/.test(rowToCheck));
-			assert.isOk(/name='Test Passing Test @pass'/.test(rowToCheck));
-			assert.isOk(/flowId=/.test(rowToCheck));
-			assert.isOk(/duration=/.test(rowToCheck));
-			assert.isOk(/]/.test(rowToCheck));
-		});
-
 		it('Test Failed is Failing', function () {
-			const rowToCheck = teamCityOutputArray[7];
+			const rowToCheck = teamCityOutputArray[4];
 			assert.isOk(/##teamcity\[testFailed/.test(rowToCheck));
-			assert.isOk(/name='Test Failing Test @fail'/.test(rowToCheck));
+			assert.isOk(/name='Failing Test @fail'/.test(rowToCheck));
 			assert.isOk(/flowId=/.test(rowToCheck));
 			assert.isOk(/duration=/.test(rowToCheck) === false);
 			assert.isOk(/details='/.test(rowToCheck));
 			assert.isOk(/AssertionError/.test(rowToCheck));
 			assert.isOk(/|n/.test(rowToCheck));
-			assert.isOk(/|failingHook.js:29:12/.test(rowToCheck));
+			assert.isOk(/|simple.js:11:11/.test(rowToCheck));
 			assert.isOk(/captureStandardOutput='true'/.test(rowToCheck));
 			assert.isOk(/]/.test(rowToCheck));
+			//check keys
+			assert.match(rowToCheck, /actual=/);
+			assert.match(rowToCheck, /expected=/);
+			//check both key and value
+			assert.match(rowToCheck, /actual='2'/);
+			assert.match(rowToCheck, /type='comparisonFailure'/);
+			assert.match(rowToCheck, /expected='1'/);
 		});
 
 		it('Failing Test Finished is OK', function () {
-			const rowToCheck = teamCityOutputArray[8];
+			const rowToCheck = teamCityOutputArray[5];
 			assert.match(rowToCheck, /##teamcity\[testFinished/);
-			assert.match(rowToCheck, /name='Test Failing Test @fail'/);
+			assert.match(rowToCheck, /name='Failing Test @fail'/);
 			assert.match(rowToCheck, /flowId=/);
 			assert.match(rowToCheck, /duration=/);
 			assert.match(rowToCheck, /]/);
 		});
 
-		it('Suite2 Finished is OK', function () {
-			const rowToCheck = teamCityOutputArray[11];
+		it('Skip Test Finished is ignored', function () {
+			const rowToCheck = teamCityOutputArray[6];
+			assert.isOk(/##teamcity\[testIgnored/.test(rowToCheck));
+			assert.isOk(/name='Skipped Test @skip'/.test(rowToCheck));
+			assert.isOk(/flowId=/.test(rowToCheck));
+			assert.isOk(/message='Skipped Test @skip'/.test(rowToCheck));
+			assert.isOk(/duration=/.test(rowToCheck) === false);
+			assert.isOk(/]/.test(rowToCheck));
+		});
+
+		it('Skip Test Finished is OK', function () {
+			const rowToCheck = teamCityOutputArray[7];
+			assert.isOk(/##teamcity\[testFinished/.test(rowToCheck));
+			assert.isOk(/name='Skipped Test @skip'/.test(rowToCheck));
+			assert.isOk(/flowId=/.test(rowToCheck));
+			assert.isNotOk(/duration=/.test(rowToCheck));
+			assert.isOk(/]/.test(rowToCheck));
+		});
+
+		it('Suite Finished is OK', function () {
+			const rowToCheck = teamCityOutputArray[8];
 			assert.isOk(/##teamcity\[testSuiteFinished/.test(rowToCheck));
-			assert.isOk(/name='Hook Test Top Describe Pass'/.test(rowToCheck));
+			assert.isOk(/name='Top Describe'/.test(rowToCheck));
 			assert.isOk(/duration=/.test(rowToCheck));
 			assert.isOk(/flowId=/.test(rowToCheck));
 			assert.isOk(/]/.test(rowToCheck));
 		});
 
 		it('Suite Root Finished is OK', function () {
-			const rowToCheck = teamCityOutputArray[11];
+			const rowToCheck = teamCityOutputArray[8];
 			assert.isOk(/##teamcity\[testSuiteFinished/.test(rowToCheck));
 			assert.isNotOk(/name='mocha.suite'/.test(rowToCheck));
 			assert.isOk(/duration=/.test(rowToCheck));
@@ -138,18 +134,19 @@ describe('Check TeamCity Output is correct with recordHookFailures option', func
 		before(function (done) {
 			const opts = {
 				env: Object.assign({
-					['RECORD_HOOK_FAILURES']: 'true'
+					['ACTUAL_VS_EXPECTED']: 'true'
 				}, process.env)
 			};
 
 			execFile(internalMochaPath, [
-				'test/example/failingHook.js',
+				'test/test_data',
 				'--reporter',
 				'lib/teamcity'
 			], opts, (err, stdout, stderr) => {
 				teamCityStdout = stdout;
 				teamCityStderr = stderr;
 				teamCityOutputArray = stdout.split('\n');
+				teamCityErrorOutputArray = stderr.split('\n');
 				logMochaOutput(stdout, stderr);
 				done();
 			});
@@ -160,15 +157,16 @@ describe('Check TeamCity Output is correct with recordHookFailures option', func
 	describe('specified with --reporter-options', function () {
 		before(function (done) {
 			execFile(internalMochaPath, [
-				path.join('test', 'example', 'failingHook.js'),
+				path.join('test', 'test_data', 'simple.js'),
 				'--reporter',
 				'lib/teamcity',
 				'--reporter-options',
-				'recordHookFailures=true'
+				'actualVsExpected=true'
 			], (err, stdout, stderr) => {
 				teamCityStdout = stdout;
 				teamCityStderr = stderr;
 				teamCityOutputArray = stdout.split('\n');
+				teamCityErrorOutputArray = stderr.split('\n');
 				logMochaOutput(stdout, stderr);
 				done();
 			});