+++ testMochaVersions.sh
@@ -4,3 +4,5 @@ npm install mocha@4
 npm run test
 npm install mocha@5
 npm run test
+npm install mocha@6
+npm run test